requirements.txt
mysqlclient>=2.0,<3.0
python-decouple>=3.6
Django>=4.0,<5.0
Pillow>=9.0,<10.0
django-imagekit
django-widget-tweaks
